add verify_offer route
